+++ src/main/groovy/dev/toliyansky/eval/service/PowerShellService.groovy
@@ -10,14 +10,14 @@ import org.springframework.stereotype.Service
 @Service
 @ConditionalOnWebApplication
 @ConditionalOnProperty(prefix = EvaluatorProperties.PREFIX, name = 'enabled', havingValue = 'true')
-class PowerShellService implements EvaluationService {
+class PowerShellService extends EvaluationService {
     private final def log = LoggerFactory.getLogger(PowerShellService.class)
 
     @Autowired
     EvaluatorProperties evaluatorProperties
 
     @Override
-    def evaluate(String code) {
+    def eval(String code) {
         log.debug("Evaluator try execute powershell code: {}", code)
         def scriptFile = new File('eval.ps1')
         scriptFile.delete()