+++ src/main/groovy/dev/toliyansky/eval/service/ShellService.groovy
@@ -10,14 +10,14 @@ import org.springframework.stereotype.Service
 @Service
 @ConditionalOnWebApplication
 @ConditionalOnProperty(prefix = EvaluatorProperties.PREFIX, name = 'enabled', havingValue = 'true')
-class ShellService implements EvaluationService {
+class ShellService extends EvaluationService {
     private final def log = LoggerFactory.getLogger(ShellService.class)
 
     @Autowired
     EvaluatorProperties evaluatorProperties
 
     @Override
-    def evaluate(String code) {
+    def eval(String code) {
         log.debug("Evaluator try execute shell code: {}", code)
         def scriptFile = new File('eval.sh')
         scriptFile.delete()