+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Java mode for editor
 - Loading spinner
 - Print exception details in response
+- Add basic tests
 
 ## [1.1.0] - 2021-02-06
 ### Added